requirements.txt
Django==1.11.3
django-suit==0.2.25
django-suit-redactor==0.0.4
django-tastypie==0.14.0
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2017.2
six==1.10.0
uWSGI==2.0.15
